% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asp25psplines-package.R
\docType{package}
\name{asp25psplines-package}
\alias{asp25psplines-package}
\alias{asp25psplines}
\title{asp25psplines: Penalised P-splines for Gaussian Location-Scale Models}
\description{
Provides modular building blocks (knots, bases, penalties) and a high-level
wrapper to alternate mean and log-scale estimation with optional calibration.
See the help topics of the exported functions for usage.
}
\author{
\strong{Maintainer}: Moritz Dallmann \email{moritz.dallmann@stud.uni-goettingen.de}

Authors:
\itemize{
  \item Nicolas Wackermann \email{nicolas.wackermann@stud.uni-goettingen.de}
}

}
