% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_spline.R
\name{fit_spline}
\alias{fit_spline}
\title{Fit a penalised B-spline smoother}
\usage{
fit_spline(x, knots)
}
\arguments{
\item{x}{Numeric covariate vector or design input.}

\item{knots}{Numeric vector of knot locations.}
}
\value{
A list with basis/design matrix and fitted coefficients.
}
\description{
Fit a penalised B-spline smoother
}
\examples{
\donttest{
x <- seq(0, 1, length.out = 50)
k <- list(
  knots = seq(0, 1, length.out = 10),
  tp    = FALSE,
  l     = 2
)
fs <- fit_spline(x = x, knots = k)
}
}
