% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_initial_values.R
\name{get_initial_values}
\alias{get_initial_values}
\title{Initial values for location-scale fitting}
\usage{
get_initial_values(X, Z, y)
}
\arguments{
\item{X}{Design matrix for the mean.}

\item{Z}{Design matrix for the log-scale.}

\item{y}{Numeric response vector.}
}
\value{
A list with initial mean and scale (or their coefficients).
}
\description{
Initial values for location-scale fitting
}
\examples{
\donttest{
# init <- get_initial_values(X, Z, y)
}
}
