% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_demo.R
\name{prep_xy_local}
\alias{prep_xy_local}
\title{prep\_xy\_local}
\usage{
prep_xy_local(x, y, sort_x = TRUE)
}
\arguments{
\item{x}{TODO: describe parameter 'x'.}

\item{y}{TODO: describe parameter 'y'.}

\item{sort\_x}{TODO: describe parameter 'sort\_x'.}
}
\value{
TODO: describe return value.
}
\description{
Internal helper function.
}
\examples{
\donttest{
# minimal example (pseudo)
# res <- prep_xy_local()
}
}
\keyword{internal}
