% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_lambda.R
\name{select_lambda}
\alias{select_lambda}
\title{select\_lambda}
\usage{
select_lambda(
  y,
  X,
  K,
  sigma_hat,
  lambda_grid,
  method = c("AIC", "BIC", "ML", "GCV"),
  ridge = 1e-08
)
}
\arguments{
\item{y}{TODO: describe parameter 'y'.}

\item{X}{TODO: describe parameter 'X'.}

\item{K}{TODO: describe parameter 'K'.}

\item{method}{TODO: describe parameter 'method'.}

\item{ridge}{TODO: describe parameter 'ridge'.}

\item{sigma\_hat}{TODO: describe parameter 'sigma\_hat'.}

\item{lambda\_grid}{TODO: describe parameter 'lambda\_grid'.}
}
\value{
TODO: describe return value.
}
\description{
Internal helper function.
}
\examples{
\donttest{
# minimal example (pseudo)
# res <- select_lambda()
}
}
\keyword{internal}
