% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_parameters.R
\name{update_parameters}
\alias{update_parameters}
\title{Update parameters for alternating location-scale fitting}
\usage{
update_parameters(
  X,
  Z,
  y,
  K_mu,
  K_sigma,
  max_iterations = 50,
  lambda_grid_mu = exp(seq(log(1e-04), log(10000), length.out = 40)),
  lambda_grid_sigma = exp(seq(log(1e-04), log(10000), length.out = 40)),
  method_mu = "BIC",
  method_sigma = "BIC",
  tolerance = 1e-05
)
}
\arguments{
\item{X}{TODO: describe parameter 'X'.}

\item{Z}{TODO: describe parameter 'Z'.}

\item{y}{TODO: describe parameter 'y'.}

\item{tolerance}{TODO: describe parameter 'tolerance'.}

\item{K\_mu}{TODO: describe parameter 'K\_mu'.}

\item{K\_sigma}{TODO: describe parameter 'K\_sigma'.}

\item{max\_iterations}{TODO: describe parameter 'max\_iterations'.}

\item{lambda\_grid\_mu}{TODO: describe parameter 'lambda\_grid\_mu'.}

\item{lambda\_grid\_sigma}{TODO: describe parameter 'lambda\_grid\_sigma'.}

\item{method\_mu}{TODO: describe parameter 'method\_mu'.}

\item{method\_sigma}{TODO: describe parameter 'method\_sigma'.}
}
\value{
TODO: describe return value.
}
\description{
Internal update engine used by high-level wrappers.
}
\examples{
\donttest{
# minimal example (pseudo)
# res <- update_parameters()
}
}
\keyword{internal}
